pyproject.toml
[project]
name = "factory-automation"
version = "0.1.0"
description = "Garment price tag factory automation system"
requires-python = ">=3.9"
dependencies = [
    # Core dependencies
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "chromadb>=0.4.0",
    "gradio>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    
    # Email processing
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    
    # ML/AI tools
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
    "litellm>=1.0.0",
    # "clip-by-openai>=1.0.0",  # Commented due to torchvision conflict
    
    # OCR
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
    
    # Data processing
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "numpy>=1.24.0",
    
    # API framework
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    
    # Utils
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.24.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "pre-commit>=3.5.0",
    "bandit[toml]>=1.7.5",
    "isort>=5.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "rag", "utils", "database", "ui", "config"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true
packages = ["factory_automation"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.bandit]
exclude_dirs = ["tests", "factory_tests", ".venv"]
skips = ["B101"]  # Skip assert_used test

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "mypy>=1.17.0",
    "types-pyyaml>=6.0.12.20250516",
]
